test(repository): repository 테스트코드 추가

+++ src/main/java/chess/repository/BoardJdbcRepository.java
@@ -44,7 +44,7 @@ public class BoardJdbcRepository implements BoardRepository {
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         boardDao.deleteById(id);
     }
